chore(menu): Rename SimpleMenu to Menu
