+++ demos/index.html
@@ -149,6 +149,13 @@
             </span>
           </a>
 
+          <a href="menu.html" role="listitem" class="mdc-list-item">
+            <span class="demo-catalog-list-icon mdc-list-item__graphic"><img src="/images/ic_menu_24px.svg" /></span>
+            <span class="mdc-list-item__text">
+              Menu
+              <span class="mdc-list-item__secondary-text">Pop over menus</span>
+            </span>
+          </a>
 
           <a href="radio.html" role="listitem" class="mdc-list-item">
             <span class="demo-catalog-list-icon mdc-list-item__graphic"><img src="/images/ic_radio_button_24px.svg" /></span>
@@ -171,14 +178,6 @@
             <span class="mdc-list-item__text">
               Select
               <span class="mdc-list-item__secondary-text">Popover selection menus</span>
-            </span>
-          </a>
-
-          <a href="simple-menu.html" role="listitem" class="mdc-list-item">
-            <span class="demo-catalog-list-icon mdc-list-item__graphic"><img src="/images/ic_menu_24px.svg" /></span>
-            <span class="mdc-list-item__text">
-              Simple Menu
-              <span class="mdc-list-item__secondary-text">Pop over menus</span>
             </span>
           </a>
