+++ demos/select.html
@@ -72,8 +72,8 @@
             <div class="mdc-select__selected-text"></div>
             <div class="mdc-select__bottom-line"></div>
           </div>
-          <div class="mdc-simple-menu mdc-select__menu">
-            <ul class="mdc-list mdc-simple-menu__items">
+          <div class="mdc-menu mdc-select__menu">
+            <ul class="mdc-list mdc-menu__items">
               <li class="mdc-list-item" role="option" tabindex="0">
                 Bread, Cereal, Rice, and Pasta
               </li>
@@ -106,8 +106,8 @@
               <div class="mdc-select__selected-text"></div>
               <div class="mdc-select__bottom-line"></div>
             </div>
-            <div class="mdc-simple-menu mdc-select__menu">
-              <ul class="mdc-list mdc-simple-menu__items">
+            <div class="mdc-menu mdc-select__menu">
+              <ul class="mdc-list mdc-menu__items">
                 <li class="mdc-list-item" role="option" id="fruit-roll-ups" tabindex="0">
                   Fruit Roll Ups
                 </li>
