+++ demos/toolbar/menu-toolbar.html
@@ -43,8 +43,8 @@
           <a href="#" class="material-icons mdc-toolbar__icon" aria-label="Print this page" alt="Print this page">print</a>
           <a href="#" class="material-icons mdc-toolbar__icon toggle" aria-label="More" alt="More">more_vert</a>
           <div class="mdc-menu-anchor">
-            <div class="mdc-simple-menu" tabindex="-1" id="demo-menu">
-              <ul class="mdc-simple-menu__items mdc-list" role="menu" aria-hidden="true" style="transform: scale(1, 1);">
+            <div class="mdc-menu" tabindex="-1" id="demo-menu">
+              <ul class="mdc-menu__items mdc-list" role="menu" aria-hidden="true" style="transform: scale(1, 1);">
                 <li class="mdc-list-item" role="menuitem" tabindex="0" style="transition-delay: 0.069s;">Back</li>
                 <li class="mdc-list-item" role="menuitem" tabindex="0" style="transition-delay: 0.124s;">Forward</li>
                 <li class="mdc-list-item" role="menuitem" tabindex="0" style="transition-delay: 0.179s;">Reload</li>
@@ -75,7 +75,7 @@
     <script>
       demoReady(function() {
         var menuEl = document.querySelector('#demo-menu');
-        var menu = new mdc.menu.MDCSimpleMenu(menuEl);
+        var menu = new mdc.menu.MDCMenu(menuEl);
         var toggle = document.querySelector('.toggle');
         toggle.addEventListener('click', function() {
           menu.open = !menu.open;
