--- packages/mdc-menu/simple/index.js
@@ -1,157 +1,0 @@
-/**
- * @license
- * Copyright 2016 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import MDCComponent from '@material/base/component';
-import {getTransformPropertyName} from '../util';
-import {MDCSimpleMenuFoundation, AnchorMargin} from './foundation';
-import {Corner, CornerBit} from './constants';
-
-/**
- * @extends MDCComponent<!MDCSimpleMenuFoundation>
- */
-class MDCSimpleMenu extends MDCComponent {
-  /** @param {...?} args */
-  constructor(...args) {
-    super(...args);
-    /** @private {!Element} */
-    this.previousFocus_;
-  }
-
-  /**
-   * @param {!Element} root
-   * @return {!MDCSimpleMenu}
-   */
-  static attachTo(root) {
-    return new MDCSimpleMenu(root);
-  }
-
-  /** @return {boolean} */
-  get open() {
-    return this.foundation_.isOpen();
-  }
-
-  /** @param {boolean} value */
-  set open(value) {
-    if (value) {
-      this.foundation_.open();
-    } else {
-      this.foundation_.close();
-    }
-  }
-
-  /** @param {{focusIndex: ?number}=} options */
-  show({focusIndex = null} = {}) {
-    this.foundation_.open({focusIndex: focusIndex});
-  }
-
-  hide() {
-    this.foundation_.close();
-  }
-
-  /**
-   * @param {Corner} corner Default anchor corner alignment of top-left
-   *     menu corner.
-   */
-  setAnchorCorner(corner) {
-    this.foundation_.setAnchorCorner(corner);
-  }
-
-  /**
-   * @param {AnchorMargin} margin
-   */
-  setAnchorMargin(margin) {
-    this.foundation_.setAnchorMargin(margin);
-  }
-
-  /**
-   * Return the item container element inside the component.
-   * @return {?Element}
-   */
-  get itemsContainer_() {
-    return this.root_.querySelector(MDCSimpleMenuFoundation.strings.ITEMS_SELECTOR);
-  }
-
-  /**
-   * Return the items within the menu. Note that this only contains the set of elements within
-   * the items container that are proper list items, and not supplemental / presentational DOM
-   * elements.
-   * @return {!Array<!Element>}
-   */
-  get items() {
-    const {itemsContainer_: itemsContainer} = this;
-    return [].slice.call(itemsContainer.querySelectorAll('.mdc-list-item[role]'));
-  }
-
-  /** @return {!MDCSimpleMenuFoundation} */
-  getDefaultFoundation() {
-    return new MDCSimpleMenuFoundation({
-      addClass: (className) => this.root_.classList.add(className),
-      removeClass: (className) => this.root_.classList.remove(className),
-      hasClass: (className) => this.root_.classList.contains(className),
-      hasNecessaryDom: () => Boolean(this.itemsContainer_),
-      getAttributeForEventTarget: (target, attributeName) => target.getAttribute(attributeName),
-      eventTargetHasClass: (target, className) => target.classList.contains(className),
-      getInnerDimensions: () => {
-        const {itemsContainer_: itemsContainer} = this;
-        return {width: itemsContainer.offsetWidth, height: itemsContainer.offsetHeight};
-      },
-      hasAnchor: () => this.root_.parentElement && this.root_.parentElement.classList.contains('mdc-menu-anchor'),
-      getAnchorDimensions: () => this.root_.parentElement.getBoundingClientRect(),
-      getWindowDimensions: () => {
-        return {width: window.innerWidth, height: window.innerHeight};
-      },
-      getNumberOfItems: () => this.items.length,
-      registerInteractionHandler: (type, handler) => this.root_.addEventListener(type, handler),
-      deregisterInteractionHandler: (type, handler) => this.root_.removeEventListener(type, handler),
-      registerBodyClickHandler: (handler) => document.body.addEventListener('click', handler),
-      deregisterBodyClickHandler: (handler) => document.body.removeEventListener('click', handler),
-      getIndexForEventTarget: (target) => this.items.indexOf(target),
-      notifySelected: (evtData) => this.emit(MDCSimpleMenuFoundation.strings.SELECTED_EVENT, {
-        index: evtData.index,
-        item: this.items[evtData.index],
-      }),
-      notifyCancel: () => this.emit(MDCSimpleMenuFoundation.strings.CANCEL_EVENT, {}),
-      saveFocus: () => {
-        this.previousFocus_ = document.activeElement;
-      },
-      restoreFocus: () => {
-        if (this.previousFocus_) {
-          this.previousFocus_.focus();
-        }
-      },
-      isFocused: () => document.activeElement === this.root_,
-      focus: () => this.root_.focus(),
-      getFocusedItemIndex: () => this.items.indexOf(document.activeElement),
-      focusItemAtIndex: (index) => this.items[index].focus(),
-      isRtl: () => getComputedStyle(this.root_).getPropertyValue('direction') === 'rtl',
-      setTransformOrigin: (origin) => {
-        this.root_.style[`${getTransformPropertyName(window)}-origin`] = origin;
-      },
-      setPosition: (position) => {
-        this.root_.style.left = 'left' in position ? position.left : null;
-        this.root_.style.right = 'right' in position ? position.right : null;
-        this.root_.style.top = 'top' in position ? position.top : null;
-        this.root_.style.bottom = 'bottom' in position ? position.bottom : null;
-      },
-      setMaxHeight: (height) => {
-        this.root_.style.maxHeight = height;
-      },
-    });
-  }
-}
-
-export {MDCSimpleMenuFoundation, MDCSimpleMenu, AnchorMargin, Corner, CornerBit};
