--- packages/mdc-menu/simple/mdc-simple-menu.scss
@@ -1,159 +1,0 @@
-//
-// Copyright 2017 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-@import "@material/animation/variables";
-@import "@material/elevation/mixins";
-@import "@material/theme/mixins";
-@import "@material/typography/mixins";
-
-$mdc-simple-menu-fade-in-duration: 0.03s;
-$mdc-simple-menu-fade-out-duration: 0.075s;
-$mdc-simple-menu-scale-duration: 0.12s;
-
-// postcss-bem-linter: define simple-menu
-.mdc-simple-menu {
-  @include mdc-elevation(2);
-
-  @include mdc-theme-dark {
-    background-color: #424242;
-  }
-
-  display: none;
-  position: absolute;
-  box-sizing: border-box;
-  min-width: 170px;
-  max-width: calc(100vw - 32px);
-  max-height: calc(100vh - 32px);
-  margin: 0;
-  padding: 0;
-  transform: scale(1);
-  transform-origin: top left;
-  border-radius: 2px;
-  background-color: white;
-  opacity: 0;
-  white-space: nowrap;
-  overflow-x: hidden;
-  overflow-y: auto;
-  will-change: transform, opacity;
-  z-index: 4;
-
-  &:focus {
-    outline: none;
-  }
-
-  &--animating-open {
-    display: inline-block;
-    overflow-y: hidden;
-    opacity: 0;
-    transform: scale(0.8);
-    transition: opacity $mdc-simple-menu-fade-in-duration linear,
-                transform $mdc-simple-menu-scale-duration $mdc-animation-deceleration-curve-timing-function;
-    > .mdc-simple-menu__items {
-      transform: scale(1.25);
-    }
-  }
-
-  &--open {
-    display: inline-block;
-    transform: scale(1);
-    opacity: 1;
-    > .mdc-simple-menu__items {
-      transform: scale(1);
-    }
-  }
-
-  &--animating-closed {
-    display: inline-block;
-    overflow-y: hidden;
-    opacity: 0;
-    transition: opacity $mdc-simple-menu-fade-out-duration linear;
-  }
-
-  &__items {
-    box-sizing: border-box;
-    overflow-x: hidden;
-    overflow-y: auto;
-    will-change: transform;
-    transform: scale(1);
-
-    > .mdc-list-item {
-      cursor: pointer;
-    }
-
-    .mdc-simple-menu--animating & {
-      overflow-y: hidden;
-    }
-  }
-
-  // stylelint-disable plugin/selector-bem-pattern
-  [dir="rtl"] & {
-    transform-origin: top right;
-  }
-  // stylelint-enable plugin/selector-bem-pattern
-
-  // stylelint-disable plugin/selector-bem-pattern
-  .mdc-list-group,
-  .mdc-list {
-    padding: 8px 0;
-  }
-
-  .mdc-list-item {
-    @include mdc-typography(subheading2);
-
-    @include mdc-theme-dark(".mdc-simple-menu", true) {
-      color: white;
-    }
-
-    position: relative;
-    outline: none;
-    color: inherit;
-    text-decoration: none;
-    user-select: none;
-  }
-
-  .mdc-list-divider {
-    @include mdc-theme-dark(".mdc-simple-menu", true) {
-      border-color: rgba(white, .12);
-    }
-  }
-
-  .mdc-list-item__graphic {
-    color: rgba(black, .54);
-
-    @include mdc-theme-dark(".mdc-simple-menu", true) {
-      color: rgba(white, .54);
-    }
-  }
-
-  .mdc-list-item[aria-disabled="true"] {
-    @include mdc-theme-prop(color, text-disabled-on-light);
-
-    @include mdc-theme-dark(".mdc-simple-menu") {
-      @include mdc-theme-prop(color, text-disabled-on-dark);
-    }
-
-    cursor: default;
-    pointer-events: none;
-
-    // TODO(#1495): Disabled menu items shouldn't be focusable, so this style shouldn't be necessary.
-    &:focus::before {
-      opacity: 0;
-    }
-  }
-  // stylelint-enable plugin/selector-bem-pattern
-}
-
-// postcss-bem-linter: end
